Responsiveness added, Fixed wording in description
Alpha Build. Responsiveness added but somehow the pictures aren't doing so great. This commit was pushed so that I have at least some responsiveness but I want to fix it as soon as possible. I do have a business day today so hopefully I can squeeze in some time to resolve it. 😊