Revised Beta Design based on class feedback
Extended the background of the dropdown and added a new gif to showcase it. Made the font for each object on the description bold so they're not the same as the description. The link to the original website now goes to a new tab. Included instructions to interaction on original website.